tentando dar push

+++ src/main/kotlin/br/com/jhonatansouza/starbuckets/model/Product.kt
@@ -1,6 +1,6 @@
 package br.com.jhonatansouza.starbuckets.model
 
-data class Product(val id: Int,
+data class Product(val id: Long,
                    val name: String,
                    val price: Double
 )
